requirements.txt
# Основные библиотеки
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic_settings>=2.0.0
loguru>=0.7.0
tenacity>=8.2.3

# Парсинг и браузерная автоматизация
playwright>=1.40.0
playwright-stealth>=1.0.5
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Асинхронность и HTTP
aiohttp>=3.9.1
aiofiles>=23.2.1
asyncio>=3.4.3
aiosignal>=1.3.1

# Работа с данными
pandas>=2.1.1
numpy>=1.25.2
matplotlib>=3.7.2

# Telegram
python-telegram-bot>=20.6

# Метрики и мониторинг
prometheus-client>=0.21.1

# Утилиты
tqdm>=4.66.1
Pillow>=10.1.0
python-dateutil>=2.8.2
pytz>=2023.3.post1
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1

# Обработка данных
python-dateutil==2.8.2

# Утилиты
xxhash==3.4.1
pytest==7.4.3
pytest-asyncio==0.23.2 